+++ benchmark/cupti_test.h
@@ -9,9 +9,10 @@
 #include <string>
 #include <stdio.h>
 #include <stdlib.h>
-#include "Metric.h"
+
+
 #include "Eval.h"
-#include "FileOp.h"
+#include "Utils.h"
 
 #ifndef EXIT_WAIVED
 #define EXIT_WAIVED 2
